Add tooltips to the idol selector
Some people may find it hard to identify which idol each image
belongs to (Especially the N's, who many people don't know as well
as the u's or Aqours girls) This patch adds a tooltip that shows
the idol's name when you hover over the image.

Note this also requires a patch to CuteForm, which I will be
submitting separately.